updated message structure

+++ src/comm/message.h
@@ -5,8 +5,8 @@
 namespace EAR {
     namespace Communication {
 	struct Message {
-	    uint32_t src;
-	    uint32_t len;
+	    uint8_t src;
+	    uint8_t len;
 	    uint8_t *buf;
 	    uint16_t crc;
 	};
